REFACTOR: Move latex utilities to analyses.rs
