+++ packages/catlog-wasm/src/lib.rs
@@ -20,7 +20,6 @@ pub mod model_presentation;
 pub mod theory;
 
 pub mod analyses;
-mod latex;
 #[allow(clippy::new_without_default)]
 #[allow(missing_docs)]
 pub mod theories;
